+++ adhocracy4/dashboard/templatetags/a4dashboard_tags.py
@@ -5,6 +5,14 @@ from adhocracy4 import phases
 register = template.Library()
 
 
+@register.filter
+def publish_value_missing(bound_field):
+    value = bound_field.value
+    if value is True or value is False:
+        return False
+    return value is None or value == ""
+
+
 @register.simple_tag
 def get_phase_name(type):
     name = phases.content[type].name