+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project (not yet) adheres to [Semantic Versioning](https://semver.org/spec/
 
 ### Unreleased
 
+- projects: add per-project `allow_guest_users` setting to control guest user
+  participation (requires `A4_ENABLE_GUEST_USERS` and django-guest-user).
+
 ## meinBerlin-v2606.3
 
 ### Added